Alternative component controller implementation (#19)
This PR provides an alternative implementation of the controller
(compared to the code currently on main).

**The main differences are:**
- **OCM Interaction:** Replacement of the `Contract` interface for
interacting with the ocm in favour of utility functions or direct calls
to the ocm (through moving generally useful utilities directly into the
library).
- **Mocks**: Reduces the usage of mocks. It primarily achieves this
through differently cut method / function abstractions and the usage of
ocm native testing utilities. This also significantly reduces the
overall necessity to provide a _kubeclient_ to general or ocm utility
functions.
The remaining reconcile logic is currently intended to be primarily
tested through `envtest` tests.

This implementation is obviously not completely finished yet. It
requires some improvement in the `envtest` tests (and also several more
of those to actually test all the important use cases). Furthermore,
several chunks of code in the main reconcile logic could probably be
aggregated into higher level utility functions.

<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->