Disable linter godox and sort list (#32)
Disable `godox` to allow `TODO:` comments and prevent a fail of the
status-check.

One could implement a smarter approach by configuring `godox` or
creating an additional status-check only checking for rather "irrelevant
problems" to display a warning. However, the functionality of `godox`
(checking lines that start with `TODO`, `FIXME` or the like) seems not
relevant and worth the time.