Documentation (#35)
#### What this PR does / why we need it:
This PR provides an initial documentation for the ocm controllers.
For now, the primary purpose of this documentation is to be able to
share knowledge with other teams (such as the landscaper team) as well
as to quickly onboard stakeholders that are willing to test our
implementation soon.

Thereby, this PR also removes some redundant status fields.

TODO: The components in the example used for the explanations do not
actually exist - so the current documentation cannot be actually played
through. This should be changed before merging. Let's care about this
working example once we have a working set of controllers.

Follow up issue:
https://github.com/open-component-model/ocm-project/issues/269

#### Which issue(s) this PR fixes:
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Co-authored-by: Jakob Möller <contact@jakob-moeller.com>