fix: automatically compress to gzip if not done so already (#54)
What this PR does / why we need it:
Resource downloads currently cant deal with non gzipped files in a way
they are consumable by the action controllers from flux without
adjustments. This fixes this by always gzipping these files so that they
can be dealt with correctly.

Example would be an OCIArtifact that contains a Tar. In this case the
download does also contain just the tar and we need to zip it along the
way.

This makes it so that all artifacts are exposing their content with gzip
by default. Other compression algorithms get converted on a best effort
basis

Which issue(s) this PR fixes: