fix: any kind of object can be a target provider as long as they implement GetTarget (#67)
#### What this PR does / why we need it:

The interface `ObjectWithTargetReference` was implement by all of the
objects as returning a `ConfigurationReference`. ConfigurationReference
in turn was using meta.NamespacedObjectKindReference inlining it. Which
is fine, however, this broke this type assert:

```go
	targetable, ok := resource.(ObjectWithTargetReference)
	if !ok {
		return nil, nil, nil, fmt.Errorf("unsupported reference type: %T", resource)
	}
```

... for all the other objects providing a target. The fix here was to
expect `*v1alpha1.ConfigurationReference` instead of
`ref.NamespacedObjectKindReference` because that's not its type during
type assertion. This is less intrusive than changing the implementation
on the objects.

I'm up for debate on that.

#### Which issue(s) this PR fixes:
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->