merge secretrefs and configrefs into single list (#81)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
This PR merges config ref and secret ref into a single list and
exchanges the implicit defaulting mechanism (implicitly inheriting the
configuration of the previous object in the pipeline - e.g. component
inherits from repository) with an explicit mechanism.

**Merging into single list**
The order in which the configuration is applied is relevant. This
becomes confusing and cumbersome with two separate lists and additional
mechanisms would be needed.

**Changing the defaulting mechanism**
1) The previous defaulting mechanism would automatically inherit the
referenced config maps if the config maps were empty and the secrets if
the secrets are empty. This does not make too much sense (but was
overseen, since originally, this was already intended as a single list).
If a user would provide all his configuration data in secrets and
therefore, doesn't specify any config maps, the mechanism would still
have inherited the config maps and potentially overwritten some of the
configs provided in the secret.
2) The new mechanism allows to only inherit parts of the configuration.
Thus, inherit the reference to config map a BUT NOT the reference to
config map b. This is useful, because the e.g. the Replication needs the
same credentials as the Component (since it needs to have access to the
component), but it does not need other parts of the configuration such
as the signing keys.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
#80