fix: component reconciler test
fixes the reconciler tests by optimizing them and cleaning them up.

To remove a potential race of the test and the control loop I had to introduce a finalizer on the component that serializes the deletion of the snapshots. (even though we are aware that envtest cannot delete with ownerrefs)