feat!: optional generic status computation (#287)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This adds a new field called `additionalStatusFields` to the Resource
Specification that replaces the old statically computed `reference`
status field. This field allows users to specify custom CEL expressions
that get evaluated at runtime in context of the OCM resource. This
allows users to dynamically resolve access specifications into
individual values, that can then be reused when integrating with other
deployers such as Flux.

Example:

A resource with an OCI Image can now look like this:

```
        spec:
          componentRef:
            name: ${component.metadata.name}
          resource:
            byReference:
              resource:
                name: helm-resource # This must match the resource name set in the OCM component version (see above)
          additionalStatusFields:
            # The additional status fields are useful for splitting the imageReference into its components, so that
            # they can be used in depending deployers
            # Example: ghcr.io/stefanprodan/charts/podinfo:6.7.1 would be 
            # registry: ghcr.io
            # repository: stefanprodan/charts/podinfo
            # reference/tag: 6.7.1
            registry: resource.access.imageReference.toOCI().registry
            repository: resource.access.imageReference.toOCI().repository
          interval: 1m
```

In this example, the additional fields `registry` and `repository` are
defined with 2 dedicated expressions. these expressions then get
evaluated in context of a serialized resource from the `compdesc`
package. the status of a resource now has an additional field called
`additional` which then contains the resolved values. This can then be
reused in a deployer. For example, in kro it would be cross-referenced
via optionals / dynamic values as such:

```
url: oci://${resourceChart.status.additional.?registry}/${resourceChart.status.additional.?repository}
```

> Note that we rely on schemaless access here!

This allows us to have a generic status computation that works with any
access and value. If there is the want for contributions on methods for
extracting status fields we now can use CEL as our expression engine to
make this dynamic

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

fix https://github.com/open-component-model/ocm-project/issues/455

---------

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>